chore: make variable type conversion explicit

+++ src/server.c
@@ -20,7 +20,7 @@ void build_http_response(const char *file_name, const char *file_ext, char *resp
 
     // Build HTTP header
     const char *mime_type = get_mime_type(file_ext);
-    char *header = (char *)malloc(BUFFER_SIZE * sizeof(char));
+    char *header = (char *)malloc((size_t)BUFFER_SIZE * sizeof(char));
     snprintf(header, BUFFER_SIZE,
              "HTTP/1.1 200 OK\r\n"
              "Content-Type: %s\r\n"
@@ -69,7 +69,7 @@ void *handle_client(void *arg)
     printf("New client thread!\n");
 
     int client_fd = *((int *)arg);
-    char *buffer = (char *)malloc(BUFFER_SIZE * sizeof(char));
+    char *buffer = (char *)malloc((size_t)BUFFER_SIZE * sizeof(char));
 
     ssize_t bytes_received = recv(client_fd, buffer, BUFFER_SIZE, 0);
 
@@ -92,7 +92,7 @@ void *handle_client(void *arg)
             strcpy(file_ext, get_file_extention(file_name));
 
             // Build HTTP Response
-            char *response = (char *)malloc(BUFFER_SIZE * 2 * sizeof(char));
+            char *response = (char *)malloc((size_t)BUFFER_SIZE * 2 * sizeof(char));
             size_t response_len;
             build_http_response(file_name, file_ext, response, &response_len);
             send(client_fd, response, response_len, 0);
